Make esc from do_look_around through ctrl-x pop back...
...to the ctrl-x screen instead of all the way back to main input.

This is more consistent with crawl's usual behaviour for cancelling
while in a nested menu or context.

This doesn't change the behaviour of the direction chooser, on the
grounds that in that case the enclosing context is the targeter, and one
opens the similar menu in direction_chooser::full_describe "on top" of
that. (I'm not super confident in this reasoning, but there it is.)